backend/requirements.txt
gunicorn==20.1.0
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
django-cors-headers==4.5.0
flake8==5.0.4
flake8-docstrings==1.7.0
django-filter==23.2
python-dotenv==1.1.1